+++ src/Query/AST/Functions/Postgresql/JsonGetPathText.php
@@ -2,36 +2,9 @@
 
 namespace Syslogic\DoctrineJsonFunctions\Query\AST\Functions\Postgresql;
 
-use Doctrine\ORM\Query\Lexer;
-use Doctrine\ORM\Query\Parser;
-
 
 class JsonGetPathText extends PostgresqlBinaryFunctionNode
 {
 	const FUNCTION_NAME = 'JSON_GET_PATH_TEXT';
 	const OPERATOR = '#>>';
-
-	/** @var \Doctrine\ORM\Query\AST\Node */
-	public $jsonData;
-
-	/** @var \Doctrine\ORM\Query\AST\Node */
-	public $jsonPath;
-
-
-	/**
-	 * @param Parser $parser
-	 */
-	public function parse(Parser $parser)
-	{
-		$parser->match(Lexer::T_IDENTIFIER);
-		$parser->match(Lexer::T_OPEN_PARENTHESIS);
-
-		$this->jsonData = $parser->StringPrimary();
-
-		$parser->match(Lexer::T_COMMA);
-
-		$this->jsonPath = $parser->StringPrimary();
-
-		$parser->match(Lexer::T_CLOSE_PARENTHESIS);
-	}
 }
