+++ src/Query/AST/Functions/Postgresql/PostgresqlBinaryFunctionNode.php
@@ -5,6 +5,8 @@ namespace Syslogic\DoctrineJsonFunctions\Query\AST\Functions\Postgresql;
 
 
 use Doctrine\ORM\Query\AST\Functions\FunctionNode;
+use Doctrine\ORM\Query\Lexer;
+use Doctrine\ORM\Query\Parser;
 use Doctrine\ORM\Query\SqlWalker;
 
 abstract class PostgresqlBinaryFunctionNode extends FunctionNode
@@ -14,6 +16,25 @@ abstract class PostgresqlBinaryFunctionNode extends FunctionNode
 
     /** @var \Doctrine\ORM\Query\AST\Node */
     protected $jsonPath;
+
+
+    /**
+     * @param Parser $parser
+     */
+    public function parse(Parser $parser)
+    {
+        $parser->match(Lexer::T_IDENTIFIER);
+        $parser->match(Lexer::T_OPEN_PARENTHESIS);
+
+        $this->jsonData = $parser->StringPrimary();
+
+        $parser->match(Lexer::T_COMMA);
+
+        $this->jsonPath = $parser->StringPrimary();
+
+        $parser->match(Lexer::T_CLOSE_PARENTHESIS);
+    }
+
 
     /**
      * @param SqlWalker $sqlWalker
